The maximum value of the field file_size in the classes Animation, Audio, Document, Video, Voice, and File can no longer be stored in a signed 32-bit integer type

+++ src/Types/Animation.php
@@ -55,9 +55,9 @@ class Animation extends BaseType
     public ?string $mime_type = null;
     /**
      * Optional. File size
-     * @var int|null
+     * @var float|null
      */
-    public ?int $file_size = null;
+    public ?float $file_size = null;
 
     /**
      * @param $key
